backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langgraph
langchain-openai
langchain_pinecone
openai
pinecone-client
python-dotenv
pypdf
tiktoken 
python-multipart
chainlit